Fix compile error

+++ lmdb/mdb.c
@@ -6858,7 +6858,7 @@ current:
 						 * Copy end of page, adjusting alignment so
 						 * compiler may copy words instead of bytes.
 						 */
-						off = (PAGEHDRSZ + data->mv_size) & -sizeof(size_t);
+						off = (size_t) ((PAGEHDRSZ + data->mv_size) & (-(signed long long) sizeof(size_t)));
 						memcpy((size_t *)((char *)np + off),
 							(size_t *)((char *)omp + off), sz - off);
 						sz = PAGEHDRSZ;
